+++ modules/apps/message-boards/message-boards-uad-test/src/testIntegration/java/com/liferay/message/boards/uad/anonymizer/test/MBMessageUADAnonymizerTest.java
@@ -9,7 +9,7 @@ import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.message.boards.model.MBMessage;
 import com.liferay.message.boards.service.MBCategoryLocalService;
 import com.liferay.message.boards.service.MBMessageLocalService;
-import com.liferay.message.boards.uad.test.MBMessageUADTestUtil;
+import com.liferay.message.boards.uad.test.util.MBMessageUADTestUtil;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
