+++ modules/apps/message-boards/message-boards-uad-test/src/testIntegration/java/com/liferay/message/boards/uad/display/test/MBMessageUADDisplayTest.java
@@ -12,8 +12,8 @@ import com.liferay.message.boards.model.MBThread;
 import com.liferay.message.boards.service.MBCategoryLocalService;
 import com.liferay.message.boards.service.MBMessageLocalService;
 import com.liferay.message.boards.service.MBThreadLocalService;
-import com.liferay.message.boards.uad.test.MBMessageUADTestUtil;
-import com.liferay.message.boards.uad.test.MBThreadUADTestUtil;
+import com.liferay.message.boards.uad.test.util.MBMessageUADTestUtil;
+import com.liferay.message.boards.uad.test.util.MBThreadUADTestUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.TestPropsValues;
